requirements.txt
tensorflow>=2.3.1
numpy>=1.18.5
matplotlib~=3.3.2
pandas~=1.1.3
scipy>=1.4.1
scikit-image~=0.17.2
tensorflow-gpu>=2.3.0
h5py>=2.10.0
opencv-python>=4.5.1.48
Pillow>=8.0.1